shuffle code a bit: put commit and putpackfile, move blob and packerjob to packer.go, get rid of unused blobs.go
