+++ snapshot/snapshot.go
@@ -1,14 +1,11 @@
 package snapshot
 
 import (
-	"bytes"
 	"crypto/rand"
-	"encoding/binary"
 	"errors"
 	"fmt"
 	"io"
 	"runtime"
-	"sync"
 	"time"
 
 	"github.com/PlakarKorp/plakar/context"
@@ -39,56 +36,6 @@ type Snapshot struct {
 
 	packerChan     chan interface{}
 	packerChanDone chan bool
-}
-
-type PackerMsg struct {
-	Timestamp time.Time
-	Type      packfile.Type
-	Checksum  objects.Checksum
-	Data      []byte
-}
-
-func packerJob(snap *Snapshot) {
-	wg := sync.WaitGroup{}
-	for i := 0; i < runtime.NumCPU(); i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			var packer *Packer
-
-			for msg := range snap.packerChan {
-				if packer == nil {
-					packer = NewPacker()
-				}
-
-				if msg, ok := msg.(*PackerMsg); !ok {
-					panic("received data with unexpected type")
-				} else {
-					snap.Logger().Trace("packer", "%x: PackerMsg(%d, %064x), dt=%s", snap.Header.GetIndexShortID(), msg.Type, msg.Checksum, time.Since(msg.Timestamp))
-					packer.AddBlob(msg.Type, msg.Checksum, msg.Data)
-				}
-
-				if packer.Size() > uint32(snap.repository.Configuration().Packfile.MaxSize) {
-					err := snap.PutPackfile(packer)
-					if err != nil {
-						panic(err)
-					}
-					packer = nil
-				}
-			}
-
-			if packer != nil {
-				err := snap.PutPackfile(packer)
-				if err != nil {
-					panic(err)
-				}
-				packer = nil
-			}
-		}()
-	}
-	wg.Wait()
-	snap.packerChanDone <- true
-	close(snap.packerChanDone)
 }
 
 func New(repo *repository.Repository) (*Snapshot, error) {
@@ -232,114 +179,6 @@ func GetSnapshot(repo *repository.Repository, Identifier objects.Checksum) (*hea
 
 func (snap *Snapshot) Repository() *repository.Repository {
 	return snap.repository
-}
-
-func (snap *Snapshot) PutPackfile(packer *Packer) error {
-
-	repo := snap.repository
-
-	serializedData, err := packer.Packfile.SerializeData()
-	if err != nil {
-		panic("could not serialize pack file data" + err.Error())
-	}
-	serializedIndex, err := packer.Packfile.SerializeIndex()
-	if err != nil {
-		panic("could not serialize pack file index" + err.Error())
-	}
-	serializedFooter, err := packer.Packfile.SerializeFooter()
-	if err != nil {
-		panic("could not serialize pack file footer" + err.Error())
-	}
-
-	encryptedIndex, err := repo.EncodeBuffer(serializedIndex)
-	if err != nil {
-		return err
-	}
-
-	encryptedFooter, err := repo.EncodeBuffer(serializedFooter)
-	if err != nil {
-		return err
-	}
-
-	encryptedFooterLength := uint8(len(encryptedFooter))
-
-	versionBytes := make([]byte, 4)
-	binary.LittleEndian.PutUint32(versionBytes, packer.Packfile.Footer.Version)
-
-	serializedPackfile := append(serializedData, encryptedIndex...)
-	serializedPackfile = append(serializedPackfile, encryptedFooter...)
-	serializedPackfile = append(serializedPackfile, versionBytes...)
-	serializedPackfile = append(serializedPackfile, byte(encryptedFooterLength))
-
-	checksum := snap.repository.Checksum(serializedPackfile)
-
-	var checksum32 objects.Checksum
-	copy(checksum32[:], checksum[:])
-
-	repo.Logger().Trace("snapshot", "%x: PutPackfile(%x, ...)", snap.Header.GetIndexShortID(), checksum32)
-	err = snap.repository.PutPackfile(checksum32, bytes.NewBuffer(serializedPackfile))
-	if err != nil {
-		panic("could not write pack file")
-	}
-
-	for _, Type := range packer.Types() {
-		for blobChecksum := range packer.Blobs[Type] {
-			for idx, blob := range packer.Packfile.Index {
-				if blob.Checksum == blobChecksum && blob.Type == Type {
-					snap.Repository().SetPackfileForBlob(Type, checksum32,
-						blobChecksum,
-						packer.Packfile.Index[idx].Offset,
-						packer.Packfile.Index[idx].Length)
-					snap.stateDelta.SetPackfileForBlob(Type, checksum32,
-						blobChecksum,
-						packer.Packfile.Index[idx].Offset,
-						packer.Packfile.Index[idx].Length)
-					break
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
-func (snap *Snapshot) Commit() error {
-
-	repo := snap.repository
-
-	serializedHdr, err := snap.Header.Serialize()
-	if err != nil {
-		return err
-	}
-
-	if kp := snap.Context().GetKeypair(); kp != nil {
-		serializedHdrChecksum := snap.repository.Checksum(serializedHdr)
-		signature := kp.Sign(serializedHdrChecksum[:])
-		if err := snap.PutBlob(packfile.TYPE_SIGNATURE, snap.Header.Identifier, signature); err != nil {
-			return err
-		}
-	}
-
-	if err := snap.PutBlob(packfile.TYPE_SNAPSHOT, snap.Header.Identifier, serializedHdr); err != nil {
-		return err
-	}
-
-	close(snap.packerChan)
-	<-snap.packerChanDone
-
-	var serializedRepositoryIndex bytes.Buffer
-	err = snap.stateDelta.SerializeStream(&serializedRepositoryIndex)
-	if err != nil {
-		snap.Logger().Warn("could not serialize repository index: %s", err)
-		return err
-	}
-	err = repo.PutState(snap.Header.Identifier, &serializedRepositoryIndex)
-	if err != nil {
-		return err
-	}
-
-	snap.Logger().Trace("snapshot", "%x: Commit()", snap.Header.GetIndexShortID())
-	return nil
 }
 
 func (snap *Snapshot) LookupObject(checksum objects.Checksum) (*objects.Object, error) {
